Revert "extract and create user done seperately"
