+++ src/routes/authRoute.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express'; // import the Router class from the express module
 
-import { generateLink, CreateUser, login } from '../controller/authController'; // import the auth controller
+import { generateLink, extractAndCreateUser, login } from '../controller/authController'; // import the auth controller
 import { validateRequestBodyToGenerateLink } from '../middleware/validation/validateRequestBodyToGenerateLink';
 import {validateLoginRequests} from '../middleware/validation/loginRequest';
 
@@ -8,7 +8,7 @@ const router = Router(); // Create an instance of the Router
 
 router.post('/generate-link', validateRequestBodyToGenerateLink, generateLink); // handle POST requests to '/generate-link' by calling the generateLink function from the authController
 
-router.get('/:id', CreateUser); // handle GET requests to '/:id' by calling the CreateUser function from the authController
+router.get('/:id', extractAndCreateUser); // handle GET requests to '/:id' by calling the extractAndCreateUser function from the authController
 
 router.post('/login', validateLoginRequests, login); // handle POST requests to '/login' by calling the login function from the authController
